% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.tukeyAdEnet}
\alias{predict.tukeyAdEnet}
\title{Predict from a fitted Tukey adaptive elastic net}
\usage{
\method{predict}{tukeyAdEnet}(object, newx, ...)
}
\arguments{
\item{object}{A `"tukeyAdEnet"` or `"tukeyAdEnetRBIC"` object, as
returned by [tukeyAdEnet()] or [tukeyAdEnetRBIC()].}

\item{newx}{Numeric predictor matrix (or an object coercible to one) with
the same number and order of columns used to fit `object`.}

\item{...}{Currently ignored.}
}
\value{
A numeric vector of predicted responses, `newx %*% coef(object)`.
}
\description{
Predict from a fitted Tukey adaptive elastic net
}
\examples{
set.seed(1)
n <- 100; p <- 10
x <- matrix(rnorm(n * p), n, p)
beta <- c(2, -2, 1.5, rep(0, p - 3))
y <- as.numeric(x \%*\% beta + rnorm(n))
fit <- tukeyAdEnet(x, y, lambda1 = 0.2, lambda2 = 0.05)
predict(fit, newx = x[1:5, ])

}
