% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyAdEnet-package.R
\docType{package}
\name{tukeyAdEnet-package}
\alias{tukeyAdEnet-package}
\title{tukeyAdEnet: Robust Adaptive Elastic Net Regression via Tukey's Biweight Loss}
\description{
Fits sparse linear regression models that are robust to outliers in the
response and in the design matrix by minimising Tukey's redescending
biweight loss subject to an adaptive elastic net penalty. Models are
fitted with a coordinate-wise proximal AdaGrad algorithm
([tukeyAdEnet()]), and regularisation parameters are selected
automatically via a robust BIC grid search ([tukeyAdEnetRBIC()]).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rkmishra1/tukey-adaptive-elasticnet}
  \item Report bugs at \url{https://github.com/rkmishra1/tukey-adaptive-elasticnet/issues}
}

}
\author{
\strong{Maintainer}: Ramakrushna Mishra \email{rkmishra.stat@gmail.com} (Indian Institute of Management Lucknow) [copyright holder]

}
\keyword{internal}
